Keep historic daily records around
